gyro-calibration/code/requirement.txt
flask
flask twilio
pyserial
panda
scipy
numpy
plotly
flask_cors